Moved color swap method to spritestore updated turn order bar to use entity's actual sprite prefab
